move some files and reorder nodes in list
