+++ backend/src/packages/chaiNNer_pytorch/pytorch/__init__.py
@@ -5,3 +5,8 @@ processing_group = pytorch_category.add_node_group("Processing")
 restoration_group = pytorch_category.add_node_group("Restoration")
 batch_processing_group = pytorch_category.add_node_group("Batch Processing")
 utility_group = pytorch_category.add_node_group("Utility")
+
+processing_group.order = [
+    "chainner:pytorch:upscale_image",
+    "chainner:pytorch:inpaint",
+]
